Block과 StaticUnit 사이의 의존 관계 변경
