+++ CPP/NativeLib/NativeLib/Block.h
@@ -1,23 +1,23 @@
 #pragma once
 
 #include "GameRule.h"
-#include "StaticUnit.h"
 #include "WorldObject.h"
+#include "BlockType.h"
 
-class Block : public WorldObject{
+class Block : public WorldObject {
 private:
 
 public:
-	StaticUnit* owner;
+	int owner_id;
 	bool is_exist;
 	eBlockType block_type;
-	Block(string _name, Transform2D _new_transform, Vector2 _new_scale, StaticUnit* _owner)
-		: WorldObject(_name, _new_transform, _new_scale), owner(_owner), is_exist(false) {
+	Block(string _name, Transform2D _new_transform, Vector2 _new_scale, int _owner_id)
+		: WorldObject(_name, _new_transform, _new_scale), owner_id(_owner_id), is_exist(false), block_type(eBlockType::NONE) {
 		isPassThrough = false;
 	}
 
 	Block(string _name, Transform2D _new_transform, Vector2 _new_scale)
-		: WorldObject(_name, _new_transform, _new_scale), is_exist(false) {
+		: WorldObject(_name, _new_transform, _new_scale), is_exist(false), block_type(eBlockType::NONE) {
 		isPassThrough = false;
 	}
