+++ CPP/NativeLib/NativeLib/BlockType.h
@@ -29,14 +29,3 @@ public:
         name[static_cast<int>(eBlockType::WOOD_FLOOR)] = "wood_floor";
     }
 };
-
-int BlockType::level[10];
-string BlockType::name[10];
-
-int BlockType::LevelOf(eBlockType type) {
-    return BlockType::level[static_cast<int>(type)];
-}
-
-string BlockType::NameOf(eBlockType type) {
-    return BlockType::name[static_cast<int>(type)];
-}
