+++ CPP/NativeLib/NativeLib/CoordinatesSystem.h
@@ -44,5 +44,6 @@ public:
 
 // To-do: 카메라 위치 반영하여 계산
 Coordinates AbsolutePositionToCoordinates(Vector2 apsolute_position);
+Coordinates AbsolutePositionToCoordinates(Transform2D absolute_position);
 
 Vector2 CoordinatesToCenterVector(Coordinates coord);
