+++ CPP/NativeLib/NativeLib/StaticUnit.h
@@ -2,27 +2,19 @@
 
 #include <vector>
 #include "Common.h"
-
-class BlockInfo {
-public:
-	int x;
-	int y;
-	int layer_level;
-
-	BlockInfo(int x, int y, int layer_level) : x(x), y(y), layer_level(layer_level) {};
-};
+#include "Block.h"
 
 class StaticUnit {
 public:
 	int id;
 	string name;
 	Rect2 ocupation_area;
-	vector<BlockInfo> blocks;
+	vector<Block*> blocks;
 
 	StaticUnit(int id, string _name, Rect2 _ocupation_area) : name(_name), ocupation_area(_ocupation_area) {}
 
-	void RegisterBlockInfo(int x, int y, int layer_level)
+	void RegisterBlock(Block* block)
 	{
-		blocks.push_back(BlockInfo(x, y, layer_level));
+		blocks.push_back(block);
 	}
 };
